monitoring/requirements.txt
psutils==3.3.14
puremagic==1.30
pypdf==6.4.0
